+++ .gitignore
@@ -1,2 +1,3 @@
 obj/
 bin/
+.vs/
